Do not trailing slash media Urls
The original Rewrite Condition would enforce a trailing slash for Urls that point to the media folder, if the file the request is for does not exists yet.

This usually includes all image FileVersion generated by Kirby when they are requested. The result is broken image Urls in the Panel or Frontend.